optmize linear hopper n iteration

+++ include/mirage/persistent_kernel/tasks/hopper/tma_3d.cuh
@@ -73,8 +73,8 @@ public:
     for (size_t i = 0; i < SMEM_REPEAT_ROW; i++) {
       for (size_t j = 0; j < SMEM_REPEAT_COL; j++) {
         int smem_offset = SMEM_STRIDE_ * j;
-        int const tma_coords_local[NDIM] = {tma_coords[0] + j * SMEM_COL,
-                                            tma_coords[1] + i * SMEM_ROW};
+        int const tma_coords_local[NDIM] = {tma_coords[0] + static_cast<int>(j * SMEM_COL),
+                                            tma_coords[1] + static_cast<int>(i * SMEM_ROW)};
 #if 0
          printf("tma_coords: %d, %d\n", tma_coords[0], tma_coords[1]);
          printf("tma_coords_local: %d, %d\n",
@@ -139,8 +139,8 @@ public:
     for (size_t i = 0; i < SMEM_REPEAT_ROW; i++) {
       for (size_t j = 0; j < SMEM_REPEAT_COL; j++) {
         int smem_offset = SMEM_STRIDE_ * j;
-        int const tma_coords_local[NDIM] = {tma_coords[0] + j * SMEM_COL,
-                                            tma_coords[1] + i * SMEM_ROW};
+        int const tma_coords_local[NDIM] = {tma_coords[0] + static_cast<int>(j * SMEM_COL),
+                                            tma_coords[1] + static_cast<int>(i * SMEM_ROW)};
         launch_tma_store_async(smem_ptr + smem_offset, tma_coords_local);
       }
     }
